initial release of new source preview page
